datadaptor: Modify Labelled Data Sets With Excel Files

An R package to modify labelled data sets with commands in Excel files. The commands in this package allow to create new variables, and modify the labels of the variables, as well as the variables themselves. The goal is to provide an easy & concise syntax, and to allow for fast systematic data entry using Excel for advanced users. The commands work on the variables inside the data.frame environment (like e.g. inside dplyr verbs), thus providing an approach that might ease the use for people without in-depth programming experience.

See also

Author

Maintainer: Urs Wilke urs.wilke@gmail.com (ORCID)

Authors:

Wolf Wilke wilke@data-connection.de